udp: update script to include share_queue and lcore mode
