+++ app/udp/ufd_server.json
@@ -1,7 +1,8 @@
 {
     "nb_udp_sockets" : "64",
     "nb_nic_queues" : "8",
-    "nic_shared_queues": true,
+    "nic_shared_queues": false,
+    "udp_lcore": false,
     "nic_queue_rate_limit_g": "2",
     "log_level" : "info",
     "fd_base" : "20000",
